+++ src/interfaces/LineChartOptions.ts
@@ -50,6 +50,21 @@ export default interface LineChartOptions {
   };
 
   /**
+   * When enabled, the line is filled with a gradient. Default is true.
+   */
+  gradient?: boolean;
+
+  /**
+   * The line width (in pixels). Default is 3.
+   */
+  lineWidth?: number;
+
+  /**
+   * When enabled, will draw the yaxis on the right side of the chart. Default is false.
+   */
+  yAxisOpposite? :boolean;
+
+  /**
    * Callback function to change the display value of the X-axis label.
    * The default is none (undefined), which displays the original value.
    * @example